Add files via upload
This works like the actual compiler and is the building blocks upon which the entire compiler. This one is the really good one. It manually sifts through each character and appends them to a string and then compares them to a set of keywords such as if, extif (extended if), and while. It then parses keywords lines like if single_condition mutable. This will helps the compiler do things like comparing dead code elimination and performing other optimizations.